requirements.txt
cx-Oracle==7.0.0
future==0.17.1
numpy==1.15.4
pandas==0.23.4
PyHive==0.6.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
sasl==0.2.1
six==1.12.0
thrift==0.11.0
thrift-sasl==0.3.0
tqdm==4.29.0
